requirements.txt
# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# GUI (optional)
PyQt5>=5.15.0

# Image processing effects
scipy>=1.10.0

# Packaging
pyinstaller>=6.0.0
